Reorganize training recipe files
