+++ examples/asr/librispeech_emformer_rnnt/lightning.py
@@ -1,10 +1,7 @@
-from argparse import ArgumentParser
 from collections import namedtuple
 import json
-import logging
 import math
 import os
-import pathlib
 from typing import List, Tuple
 
 from fairseq.data import Dictionary
@@ -15,11 +12,8 @@ import torchaudio
 import torchaudio.functional as F
 from torchaudio.prototype.rnnt import emformer_rnnt_base
 from torchaudio.prototype.rnnt_decoder import Hypothesis, RNNTBeamSearch
-from pytorch_lightning import LightningModule, Trainer
-from pytorch_lightning.callbacks import ModelCheckpoint
+from pytorch_lightning import LightningModule
 
-
-logger = logging.getLogger()
 
 Batch = namedtuple(
     "Batch", ["features", "feature_lengths", "targets", "target_lengths"]
@@ -160,6 +154,34 @@ class WarmupLR(torch.optim.lr_scheduler._LRScheduler):
             (min(1.0, self._step_count / self.warmup_updates)) * base_lr
             for base_lr in self.base_lrs
         ]
+
+
+def post_process_hypos(
+    hypos: List[Hypothesis], sp_model: spm.SentencePieceProcessor, tgt_dict: Dictionary
+) -> List[Tuple[str, float, List[int], List[int]]]:
+    post_process_remove_list = [
+        sp_model.unk_id(),
+        sp_model.eos_id(),
+        sp_model.pad_id(),
+    ]
+    hypos_str = [
+        tgt_dict.string(
+            [
+                token_index
+                for token_index in h.tokens[1:]
+                if token_index not in post_process_remove_list
+            ]
+        )
+        for h in hypos
+    ]
+    hypos_str = [sp_model.DecodePieces(s.split()) for s in hypos_str]
+    hypos_ali = [h.alignment[1:] for h in hypos]
+    hypos_ids = [h.tokens[1:] for h in hypos]
+    hypos_score = [[math.exp(h.score)] for h in hypos]
+
+    nbest_batch = list(zip(hypos_str, hypos_score, hypos_ali, hypos_ids))
+
+    return nbest_batch
 
 
 class RNNTModule(LightningModule):
@@ -369,167 +391,3 @@ class RNNTModule(LightningModule):
             dataset, batch_size=1, collate_fn=self._test_collate_fn
         )
         return dataloader
-
-
-def compute_word_level_distance(seq1, seq2):
-    return torchaudio.functional.edit_distance(
-        seq1.lower().split(), seq2.lower().split()
-    )
-
-
-def post_process_hypos(
-    hypos: List[Hypothesis], sp_model: spm.SentencePieceProcessor, tgt_dict: Dictionary
-) -> List[Tuple[str, float, List[int], List[int]]]:
-    post_process_remove_list = [
-        sp_model.unk_id(),
-        sp_model.eos_id(),
-        sp_model.pad_id(),
-    ]
-    hypos_str = [
-        tgt_dict.string(
-            [
-                token_index
-                for token_index in h.tokens[1:]
-                if token_index not in post_process_remove_list
-            ]
-        )
-        for h in hypos
-    ]
-    hypos_str = [sp_model.DecodePieces(s.split()) for s in hypos_str]
-    hypos_ali = [h.alignment[1:] for h in hypos]
-    hypos_ids = [h.tokens[1:] for h in hypos]
-    hypos_score = [[math.exp(h.score)] for h in hypos]
-
-    nbest_batch = list(zip(hypos_str, hypos_score, hypos_ali, hypos_ids))
-
-    return nbest_batch
-
-
-def run_eval(args):
-    model = (
-        RNNTModule.load_from_checkpoint(
-            args.checkpoint_path,
-            librispeech_path=args.librispeech_path,
-            sp_model_path=args.sp_model_path,
-            tgt_dict_path=args.tgt_dict_path,
-            global_stats_path=args.global_stats_path,
-        )
-        .eval()
-        .to(device="cuda")
-    )
-    total_edit_distance = 0
-    total_length = 0
-    dataloader = model.test_dataloader()
-    with torch.no_grad():
-        for idx, (batch, sample) in enumerate(dataloader):
-            actual = sample[0][2]
-            predicted = model(batch)
-            total_edit_distance += compute_word_level_distance(actual, predicted)
-            total_length += len(actual.split())
-            if idx % 100 == 0:
-                logger.info(
-                    f"Processed elem {idx}; WER: {total_edit_distance / total_length}"
-                )
-    logger.info(f"Final WER: {total_edit_distance / total_length}")
-
-
-def run_train(args):
-    checkpoint_dir = args.exp_dir / "checkpoints"
-    checkpoint = ModelCheckpoint(
-        checkpoint_dir,
-        monitor="Losses/val_loss",
-        mode="min",
-        save_top_k=10,
-        save_weights_only=True,
-        verbose=True,
-    )
-    train_checkpoint = ModelCheckpoint(
-        checkpoint_dir,
-        monitor="Losses/train_loss",
-        mode="min",
-        save_top_k=10,
-        save_weights_only=True,
-        verbose=True,
-    )
-    callbacks = [
-        checkpoint,
-        train_checkpoint,
-    ]
-    trainer = Trainer(
-        default_root_dir=args.exp_dir,
-        max_epochs=args.epochs,
-        num_nodes=args.num_nodes,
-        gpus=args.gpus,
-        accelerator="gpu",
-        strategy="ddp",
-        gradient_clip_val=10.0,
-        callbacks=callbacks,
-    )
-
-    model = RNNTModule(
-        librispeech_path=args.librispeech_path,
-        sp_model_path=args.sp_model_path,
-        tgt_dict_path=args.tgt_dict_path,
-        global_stats_path=args.global_stats_path,
-    )
-    trainer.fit(model)
-
-
-def cli_main():
-    parser = ArgumentParser()
-    parser.add_argument(
-        "--eval", action="store_true", default=False, help="Run in eval mode.",
-    )
-    parser.add_argument(
-        "--exp_dir",
-        default=pathlib.Path("./exp"),
-        type=pathlib.Path,
-        help="Directory to save checkpoints and logs to. (Default: './exp')",
-    )
-    parser.add_argument(
-        "--checkpoint_path",
-        type=pathlib.Path,
-        help="Path to checkpoint to use for evaluation.",
-    )
-    parser.add_argument(
-        "--global_stats_path",
-        type=str,
-        help="Path to JSON file containing feature means and stddevs.",
-    )
-    parser.add_argument(
-        "--librispeech_path", type=str, help="Path to LibriSpeech datasets.",
-    )
-    parser.add_argument(
-        "--sp_model_path", type=str, help="Path to SentencePiece model.",
-    )
-    parser.add_argument(
-        "--tgt_dict_path", type=str, help="Path to fairseq token dictionary.",
-    )
-    parser.add_argument(
-        "--num_nodes",
-        default=4,
-        type=int,
-        help="Number of nodes to use for training. (Default: 4)",
-    )
-    parser.add_argument(
-        "--gpus",
-        default=8,
-        type=int,
-        help="Number of GPUs per node to use for training. (Default: 8)",
-    )
-    parser.add_argument(
-        "--epochs",
-        default=120,
-        type=int,
-        help="Number of epochs to train for. (Default: 120)",
-    )
-    args = parser.parse_args()
-
-    if args.eval:
-        run_eval(args)
-    else:
-        run_train(args)
-
-
-if __name__ == "__main__":
-    cli_main()
